refactor(gnb): 훅, 상태를 컴포넌트 내부로 리팩토링
