+++ src/pages/SharedPage.jsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import { getFolder, getUser } from '../apis/api';
+import { getFolder } from '../apis/api';
 
 import Gnb from '../components/Gnb';
 import Header from '../Header';
@@ -7,7 +7,6 @@ import Folder from '../Folder';
 import Footer from '../components/Footer';
 
 const SharedPage = () => {
-  const [users, setUsers] = useState({});
   const [folders, setFolders] = useState({});
 
   useEffect(() => {
@@ -20,21 +19,11 @@ const SharedPage = () => {
     fetchFolder();
   }, []);
 
-  useEffect(() => {
-    const fetchUser = async () => {
-      const userInfo = await getUser();
-      
-      setUsers(userInfo);
-    };
-
-    fetchUser();
-  }, []);
-
   return (
     <>
-      <Gnb items={users}/>
-      <Header items={folders}/>
-      <Folder items={folders}/>
+      <Gnb />
+      <Header items={folders} />
+      <Folder items={folders} />
       <Footer/>
     </>
   )
